webhooks: merge messages_handler.go into message.go
MessagesHandler, MediaHandler, and InteractiveHandler now live alongside
MessageType constants, MessageHandler interface, and all On*/Set*
registration methods in message.go — a single file for all message
concerns.
